+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.4] - 2026-01-21
+
+### Fixed
+- Fix fountain-js parser integration (was using wrong import/API)
+- Validate and export commands now work correctly with Fountain files
+
+### Added
+- Test for fountain file validation
+
 ## [1.0.3] - 2026-01-21
 
 ### Added
@@ -44,6 +53,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `wtfb export-html` - Export screenplay to HTML
 - `wtfb info` - Show CLI information
 
+[1.0.4]: https://github.com/bybren-llc/wtfb-packages/compare/v1.0.3...v1.0.4
 [1.0.3]: https://github.com/bybren-llc/wtfb-packages/compare/v1.0.2...v1.0.3
 [1.0.2]: https://github.com/bybren-llc/wtfb-packages/compare/v1.0.1...v1.0.2
 [1.0.1]: https://github.com/bybren-llc/wtfb-packages/compare/v1.0.0...v1.0.1